[UWP] Move RunAsync out of VFS implementation
