+++ libretro-common/vfs/vfs_implementation_uwp.cpp
@@ -55,62 +55,10 @@ using namespace Windows::Storage::FileProperties;
 #include <string/stdstring.h>
 #include <retro_environment.h>
 #include <uwp/uwp_func.h>
+#include <uwp/uwp_async.h>
 
 namespace
 {
-	/* Dear Microsoft
-	 * I really appreciate all the effort you took to not provide any
-	 * synchronous file APIs and block all attempts to synchronously
-	 * wait for the results of async tasks for "smooth user experience",
-	 * but I'm not going to run and rewrite all RetroArch cores for
-	 * async I/O. I hope you like this hack I made instead.
-	 */
-	template<typename T>
-	T RunAsync(std::function<concurrency::task<T>()> func)
-	{
-		volatile bool finished = false;
-		Platform::Exception^ exception = nullptr;
-		T result;
-
-		func().then([&finished, &exception, &result](concurrency::task<T> t) {
-			try
-			{
-				result = t.get();
-			}
-			catch (Platform::Exception^ exception_)
-			{
-				exception = exception_;
-			}
-			finished = true;
-		});
-
-		/* Don't stall the UI thread - prevents a deadlock */
-		Windows::UI::Core::CoreWindow^ corewindow = Windows::UI::Core::CoreWindow::GetForCurrentThread();
-		while (!finished)
-		{
-			if (corewindow) {
-				corewindow->Dispatcher->ProcessEvents(Windows::UI::Core::CoreProcessEventsOption::ProcessAllIfPresent);
-			}
-		}
-
-		if (exception != nullptr)
-			throw exception;
-		return result;
-	}
-
-	template<typename T>
-	T RunAsyncAndCatchErrors(std::function<concurrency::task<T>()> func, T valueOnError)
-	{
-		try
-		{
-			return RunAsync<T>(func);
-		}
-		catch (Platform::Exception^ e)
-		{
-			return valueOnError;
-		}
-	}
-
 	void windowsize_path(wchar_t* path)
 	{
 		/* UWP deals with paths containing / instead of \ way worse than normal Windows */
