+++ pkg/msvc-uwp/RetroArch-msvc2017-UWP/RetroArch-msvc2017-UWP.vcxproj.filters
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
     <Filter Include="uwp">
@@ -27,6 +27,9 @@
       <Filter>uwp</Filter>
     </ClInclude>
     <ClInclude Include="..\..\..\uwp\uwp_main.h">
+      <Filter>uwp</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\uwp\uwp_async.h">
       <Filter>uwp</Filter>
     </ClInclude>
   </ItemGroup>
